requirements.txt
google-generativeai
rapidfuzz
flask
flask-cors
twilio